+++ workout-score/App.tsx
@@ -1,35 +1,22 @@
-import { StyleSheet } from 'react-native';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { NavigationContainer } from '@react-navigation/native';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { HomeScreen } from './src/screens/HomeScreen';
+import { SettingsScreen } from './src/screens/SettingsScreen';
+import { WorkoutScreen } from './src/screens/WorkoutScreen';
+
+const Tab = createBottomTabNavigator();
 
 export default function App() {
   return (
     <SafeAreaProvider>
-      <HomeScreen />
+      <NavigationContainer>
+        <Tab.Navigator initialRouteName="Home">
+          <Tab.Screen name="Home" component={HomeScreen} />
+          <Tab.Screen name="Workout" component={WorkoutScreen} />
+          <Tab.Screen name="Settings" component={SettingsScreen} />
+        </Tab.Navigator>
+      </NavigationContainer>
     </SafeAreaProvider>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
-
-
-
-// <SafeAreaProvider>
-      // <NavigationContainer>
-      //   <Stack.Navigator>
-      //     <Stack.Screen name='Home' component={HomeScreen} />
-      //     <Stack.Screen name="Setting" component={SettingScreen} />
-      //   </Stack.Navigator>
-      // </NavigationContainer>
-//       <StatusBar barStyle="dark-content" />
-//       {/* <View style={styles.container}>
-//         <Text>Open up App.tsx to start working on your app!</Text>
-//       </View> */}
-//     </SafeAreaProvider>
