+++ workout-score/src/screens/SettingsScreen.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+export const SettingsScreen = () => {
+  return (
+    <View style={styles.container}>
+      <Text>SettingScreen</Text>
+    </View>
+  )
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
